+++ hexagonit/virtualgallery/docs/HISTORY.rst
@@ -4,7 +4,8 @@ Changelog
 1.0b3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- JSON schema validation.
+  [zupo]
 
 
 1.0b2 (2011-08-04)